Configure_package_json_for_a_node_application
Your task
Create a package.json configuration and assign it to the configuration constant so it can be tested.

To pass the kata you need only the least that a real package file would require.

Notes
the kata version validation doesn't handle all cases and isn't 100% accurate, so don't try to be too fancy, stick to with the simple